final commit

+++ includes/manage_prints_functions.php
@@ -30,43 +30,45 @@ if(!empty($_POST['printfile'])) {
 * 
 */
 if (isset($_POST['start_button'])) {		
-	if ($next_print = get_next_print($mysqli)) {																			// Sofern es überhaupt einen nächsten Druck gibt
-		if (!empty($next_print)) {
-			$ulti->put("printer/led", json_encode(array('brightness' => '100', 'hue' => '100', 'saturation' => '100')));	// Ändere die Farbe der LEDs im Drucker auf Grün
-			if ($stmt = $mysqli->prepare("UPDATE current_state SET `value` = ? WHERE `key` = 'current_print_id'")) {		// Setze die Informationen über den aktuellen Druck in der DB
-				$stmt->bind_param('s', $next_print[8]);
-				if (!$stmt->execute()) {
-					header('Location: ../error.php?err=Druck Fehler: UPDATE current_print_id');
-				}
-				if ($stmt = $mysqli->prepare("UPDATE current_state SET `value` = ? WHERE `key` = 'current_print_file'")) {
-					$stmt->bind_param('s', $next_print[3]);
+	if (get_status($ulti) == "im Leerlauf") {
+		if ($next_print = get_next_print($mysqli)) {																			// Sofern es überhaupt einen nächsten Druck gibt
+			if (!empty($next_print)) {
+				$ulti->put("printer/led", json_encode(array('brightness' => '100', 'hue' => '100', 'saturation' => '100')));	// Ändere die Farbe der LEDs im Drucker auf Grün
+				if ($stmt = $mysqli->prepare("UPDATE current_state SET `value` = ? WHERE `key` = 'current_print_id'")) {		// Setze die Informationen über den aktuellen Druck in der DB
+					$stmt->bind_param('s', $next_print[8]);
 					if (!$stmt->execute()) {
-						header('Location: ../error.php?err=Druck Fehler: UPDATE current_print_file');
+						header('Location: ../error.php?err=Druck Fehler: UPDATE current_print_id');
 					}
-					if ($stmt = $mysqli->prepare("UPDATE current_state SET `value` = ? WHERE `key` = 'current_print_owner'")) {
-						$stmt->bind_param('s', $next_print[0]);
+					if ($stmt = $mysqli->prepare("UPDATE current_state SET `value` = ? WHERE `key` = 'current_print_file'")) {
+						$stmt->bind_param('s', $next_print[3]);
 						if (!$stmt->execute()) {
-							header('Location: ../error.php?err=Druck Fehler: UPDATE current_print_owner');
+							header('Location: ../error.php?err=Druck Fehler: UPDATE current_print_file');
 						}
-						if ($stmt = $mysqli->prepare("DELETE FROM to_print WHERE to_print.id = ?")) {					// Entferne den Druck aus der Warteschlange
-							$stmt->bind_param('i', $next_print[8]);
+						if ($stmt = $mysqli->prepare("UPDATE current_state SET `value` = ? WHERE `key` = 'current_print_owner'")) {
+							$stmt->bind_param('s', $next_print[0]);
 							if (!$stmt->execute()) {
-								header('Location: ../error.php?err=Druck Fehler: DELETE to_print');
+								header('Location: ../error.php?err=Druck Fehler: UPDATE current_print_owner');
 							}
-							if($ulti->post_file($next_print[3])) {														// Wenn der Druck korrekt gestartet wird, erstelle Emails in DB
-								$expected_finish = time() + (floatval($next_print[5]) * 1.03) + 60 * 2;					// 60 * 2 simuliert die 2 Minuten die der Drucker ca. benötigt um Aufzuheizen, 3% für die Abweichungen
-								new_mail($mysqli, $next_print[8], time(), $next_print[0], $next_print[3], " durch den Benutzer " . $ulti->get_user()->get_user_email() . " gestartet, erwartete Fertigstellung: ca. " . date("H:i \U\h\\r\, \a\m d.m", $expected_finish));
-								new_mail($mysqli, $next_print[8], $expected_finish, $next_print[0], $next_print[3], "fertiggestellt");
-								if (!empty($over_next_print = get_next_print($mysqli))) {								// Wenn es einen übernächsten Druck gibt, erstelle für den Besitzer eine Benachrichtigung
-									new_mail($mysqli, $next_print[8], $expected_finish, $over_next_print[0], $over_next_print[3], "druckbereit");
+							if ($stmt = $mysqli->prepare("DELETE FROM to_print WHERE to_print.id = ?")) {					// Entferne den Druck aus der Warteschlange
+								$stmt->bind_param('i', $next_print[8]);
+								if (!$stmt->execute()) {
+									header('Location: ../error.php?err=Druck Fehler: DELETE to_print');
+								}
+								if($ulti->post_file($next_print[3])) {														// Wenn der Druck korrekt gestartet wird, erstelle Emails in DB
+									$expected_finish = time() + (floatval($next_print[5]) * 1.03) + 60 * 2;					// 60 * 2 simuliert die 2 Minuten die der Drucker ca. benötigt um Aufzuheizen, 3% für die Abweichungen
+									new_mail($mysqli, $next_print[8], time(), $next_print[0], $next_print[3], " durch den Benutzer " . $ulti->get_user()->get_user_email() . " gestartet, erwartete Fertigstellung: ca. " . date("H:i \U\h\\r\, \a\m d.m", $expected_finish));
+									new_mail($mysqli, $next_print[8], $expected_finish, $next_print[0], $next_print[3], "fertiggestellt");
+									if (!empty($over_next_print = get_next_print($mysqli))) {								// Wenn es einen übernächsten Druck gibt, erstelle für den Besitzer eine Benachrichtigung
+										new_mail($mysqli, $next_print[8], $expected_finish, $over_next_print[0], $over_next_print[3], "druckbereit");
+									}
 								}
 							}
 						}
 					}
 				}
-			}
-		} else echo "<div id='response'>Kein Druckauftrag in der Warteschlange.</div>";
-	}
+			} else echo "<div id='response'>Fehler: Kein Druckauftrag in der Warteschlange.</div>";
+		}
+	} else echo "<div id='response'>Fehler: Der Drucker ist nicht im Leerlauf</div>";
 }
 
 /**
